+++ source/public/form.html
@@ -6,9 +6,9 @@
   <script src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.7.3/handlebars.min.js"
           integrity="sha256-/PJBs6QWvXijOFIX04kZpLb6ZtSQckdOIavLWKKOgXU=" crossorigin="anonymous" defer></script>
   <script src="scripts/index.js" type="module"></script>
-  <script src="scripts/todo-store.js" type="module"></script>
+  <script src="scripts/services/data/todo-store.js" type="module"></script>
 
-  <link rel="stylesheet" href="styles/index.css">
+  <link rel="stylesheet" href="styles/main.css">
 </head>
 <body>
 <header>
